# util/swissmap: Go implementation of Swiss Tables