@petermattis 2024-02-04
This was how I spent my time on the 20 hours of plane flight to India. I don't have an immediate usage for this map, but we should probably replace Pebble's block cache robinhood map with this one. I suspect there might be a few prominent maps within CRDB that could be replaced with `swissmap.M`. This likely deserves to be in a separate repository from CRDB (i.e. github.com/cockroachdb/swissmap).