@RaduBerinde 2024-02-05
*[`pkg/util/swissmap/swissmap.go` line 449 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/118742#-NpoJMCD15eB7n8E6FEf:-Npu39Y0D8yloKbHXkn0:b-r3qfpk) ([raw file](https://github.com/cockroachdb/cockroach/blob/41364cb3bbe17c9ea1336aedb4fcb88f8110aa06/pkg/util/swissmap/swissmap.go#L449)):*
> When we rehash in place to clear out tombstones we won't be doubling the capacity of the buffer.

How are we not? The only call to `rehash` (other than the initial one) is `rehash(2*m.capacity + 1)`


<!-- Sent from Reviewable.io -->