@RaduBerinde 2024-02-05
*[`pkg/util/swissmap/swissmap.go` line 449 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/118742#-NpoJMCD15eB7n8E6FEf:-Npu3_aF46D_h2l8a-eH:bblobme) ([raw file](https://github.com/cockroachdb/cockroach/blob/41364cb3bbe17c9ea1336aedb4fcb88f8110aa06/pkg/util/swissmap/swissmap.go#L449)):*
<details><summary><i>Previously, RaduBerinde wrote…</i></summary><blockquote>

> When we rehash in place to clear out tombstones we won't be doubling the capacity of the buffer.

How are we not? The only call to `rehash` (other than the initial one) is `rehash(2*m.capacity + 1)`
</blockquote></details>

Oh, do you mean when we will implement rehashing in place? Well we don't do that right now. The code as it is now always doubles the capacity when we rehash. All I'm proposing is instead of calling `rehash(2*m.capacity + 1)` to call `rehash(roundToPowerOfTwo(m.used*2)-1)`


<!-- Sent from Reviewable.io -->